requirements.txt
pytest~=8.4.1
pytest-xdist
pytest-cov
pytest-timeout
allure-pytest
allure-python-commons
flake8
black
isort
mypy
bandit